requirements.txt
fastapi
torch
torchvision
ultralytics
simple-lama-inpainting
opencv-python-headless
Pillow
numpy
aiofiles
python-multipart
moviepy
imageio[ffmpeg]
uvicorn[standard]
